requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# AI Services
google-generativeai==0.8.3

# Speech Recognition
SpeechRecognition==3.10.0
PyAudio==0.2.14

# Text-to-Speech
edge-tts==6.1.9

# Production Server
gunicorn==21.2.0